backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
neo4j==5.14.1
faiss-cpu==1.7.4
transformers==4.35.2
torch==2.8.0
numpy==1.26.4
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
